refactor: [#1195] extract core::authentication::handler::KeysHandler
