+++ src/servers/apis/v1/context/auth_key/handlers.rs
@@ -12,7 +12,8 @@ use super::responses::{
     auth_key_response, failed_to_delete_key_response, failed_to_generate_key_response, failed_to_reload_keys_response,
     invalid_auth_key_duration_response, invalid_auth_key_response,
 };
-use crate::core::authentication::{AddKeyRequest, Key};
+use crate::core::authentication::handler::AddKeyRequest;
+use crate::core::authentication::Key;
 use crate::core::Tracker;
 use crate::servers::apis::v1::context::auth_key::resources::AuthKey;
 use crate::servers::apis::v1::responses::{invalid_auth_key_param_response, ok_response};
